Update my keymap on Pila87 and some minor changes
